Add the most basic validation to services

+++ src/main/java/by/clever/bank/service/impl/AccountServiceImpl.java
@@ -6,6 +6,7 @@ import by.clever.bank.dao.DAOFactory;
 import by.clever.bank.dao.exception.DAOException;
 import by.clever.bank.service.AccountService;
 import by.clever.bank.service.exception.ServiceException;
+import by.clever.bank.service.exception.ServiceValidationException;
 import by.clever.bank.transactionmanager.AccountTransactionManager;
 import by.clever.bank.transactionmanager.TransactionManagerFactory;
 import by.clever.bank.transactionmanager.exception.TransactionManagerException;
@@ -21,8 +22,12 @@ public class AccountServiceImpl implements AccountService {
     public boolean putMoneyToAccount(BigDecimal amount, String accountNumber) throws ServiceException {
 
         try {
-            return transactionManager.putMoneyToAccount(amount, accountNumber);
-        } catch (TransactionManagerException e) {
+            if (amount != null && accountNumber != null) {
+                return transactionManager.putMoneyToAccount(amount, accountNumber);
+            } else {
+                throw new ServiceValidationException("Null arguments are not accepted");
+            }
+        } catch (TransactionManagerException | ServiceValidationException e) {
             throw new ServiceException(e);
         }
     }
@@ -30,8 +35,12 @@ public class AccountServiceImpl implements AccountService {
     @Override
     public boolean withdrawMoneyFromAccount(BigDecimal amount, String accountNumber) throws ServiceException {
         try {
-            return transactionManager.withdrawMoneyFromAccount(amount, accountNumber);
-        } catch (TransactionManagerException e) {
+            if (amount != null && accountNumber != null) {
+                return transactionManager.withdrawMoneyFromAccount(amount, accountNumber);
+            } else {
+                throw new ServiceValidationException("Null arguments are not accepted");
+            }
+        } catch (TransactionManagerException | ServiceValidationException e) {
             throw new ServiceException(e);
         }
     }
@@ -39,8 +48,12 @@ public class AccountServiceImpl implements AccountService {
     @Override
     public boolean transferMoneyBetweenAccounts(BigDecimal amount, String senderAccountNumber, String receiverAccountNumber) throws ServiceException {
         try {
-            return transactionManager.transferMoneyBetweenAccounts(amount, senderAccountNumber, receiverAccountNumber);
-        } catch (TransactionManagerException e) {
+            if (amount != null && senderAccountNumber != null && receiverAccountNumber != null) {
+                return transactionManager.transferMoneyBetweenAccounts(amount, senderAccountNumber, receiverAccountNumber);
+            } else {
+                throw new ServiceValidationException("Null arguments are not accepted");
+            }
+        } catch (TransactionManagerException | ServiceValidationException e) {
             throw new ServiceException(e);
         }
     }
@@ -48,8 +61,12 @@ public class AccountServiceImpl implements AccountService {
     @Override
     public boolean createAccount(Account account, String bankName, int userID) throws ServiceException {
         try {
-            return accountDAO.createAccount(account, bankName, userID);
-        } catch (DAOException e) {
+            if (account != null && bankName != null && userID > 0) {
+                return accountDAO.createAccount(account, bankName, userID);
+            } else {
+                throw new ServiceValidationException("Null and negative or zero arguments are not accepted");
+            }
+        } catch (DAOException | ServiceValidationException e) {
             throw new ServiceException(e);
         }
     }
@@ -57,8 +74,12 @@ public class AccountServiceImpl implements AccountService {
     @Override
     public boolean updateAccount(Account oldAccount, Account newAccount) throws ServiceException {
         try {
-            return accountDAO.updateAccount(oldAccount, newAccount);
-        } catch (DAOException e) {
+            if (oldAccount != null && newAccount != null) {
+                return accountDAO.updateAccount(oldAccount, newAccount);
+            } else {
+                throw new ServiceValidationException("Null and negative or zero arguments are not accepted");
+            }
+        } catch (DAOException | ServiceValidationException e) {
             throw new ServiceException(e);
         }
     }
@@ -66,8 +87,12 @@ public class AccountServiceImpl implements AccountService {
     @Override
     public Account readAccount(int accountID) throws ServiceException {
         try {
-            return accountDAO.readAccount(accountID);
-        } catch (DAOException e) {
+            if (accountID > 0) {
+                return accountDAO.readAccount(accountID);
+            } else {
+                throw new ServiceValidationException("Negative or zero arguments are not accepted");
+            }
+        } catch (DAOException | ServiceValidationException e) {
             throw new ServiceException(e);
         }
     }
@@ -75,8 +100,12 @@ public class AccountServiceImpl implements AccountService {
     @Override
     public boolean deleteAccount(int accountID) throws ServiceException {
         try {
-            return accountDAO.deleteAccount(accountID);
-        } catch (DAOException e) {
+            if (accountID > 0) {
+                return accountDAO.deleteAccount(accountID);
+            } else {
+                throw new ServiceValidationException("Negative or zero arguments are not accepted");
+            }
+        } catch (DAOException | ServiceValidationException e) {
             throw new ServiceException(e);
         }
     }
