+++ src/main/java/by/clever/bank/service/impl/TransactionServiceImpl.java
@@ -6,6 +6,7 @@ import by.clever.bank.dao.TransactionDAO;
 import by.clever.bank.dao.exception.DAOException;
 import by.clever.bank.service.TransactionService;
 import by.clever.bank.service.exception.ServiceException;
+import by.clever.bank.service.exception.ServiceValidationException;
 
 public class TransactionServiceImpl implements TransactionService {
 
@@ -14,8 +15,12 @@ public class TransactionServiceImpl implements TransactionService {
     @Override
     public boolean createTransaction(Transaction transaction, int accountID) throws ServiceException {
         try {
-            return transactionDAO.createTransaction(transaction, accountID);
-        } catch (DAOException e) {
+            if (transaction != null && accountID > 0) {
+                return transactionDAO.createTransaction(transaction, accountID);
+            } else {
+                throw new ServiceValidationException("Null and negative or zero arguments are not accepted");
+            }
+        } catch (DAOException | ServiceValidationException e) {
             throw new ServiceException(e);
         }
     }
@@ -23,8 +28,12 @@ public class TransactionServiceImpl implements TransactionService {
     @Override
     public boolean updateTransaction(Transaction transaction, int transactionID) throws ServiceException {
         try {
-            return transactionDAO.updateTransaction(transaction, transactionID);
-        } catch (DAOException e) {
+            if (transaction != null && transactionID > 0) {
+                return transactionDAO.updateTransaction(transaction, transactionID);
+            } else {
+                throw new ServiceValidationException("Null and negative or zero arguments are not accepted");
+            }
+        } catch (DAOException | ServiceValidationException e) {
             throw new ServiceException(e);
         }
     }
@@ -32,8 +41,12 @@ public class TransactionServiceImpl implements TransactionService {
     @Override
     public Transaction readTransaction(int transactionID) throws ServiceException {
         try {
-            return transactionDAO.readTransaction(transactionID);
-        } catch (DAOException e) {
+            if (transactionID > 0) {
+                return transactionDAO.readTransaction(transactionID);
+            } else {
+                throw new ServiceValidationException("Negative or zero arguments are not accepted");
+            }
+        } catch (DAOException | ServiceValidationException e) {
             throw new ServiceException(e);
         }
     }
@@ -41,8 +54,12 @@ public class TransactionServiceImpl implements TransactionService {
     @Override
     public boolean deleteTransaction(int transactionID) throws ServiceException {
         try {
-            return transactionDAO.deleteTransaction(transactionID);
-        } catch (DAOException e) {
+            if (transactionID > 0) {
+                return transactionDAO.deleteTransaction(transactionID);
+            } else {
+                throw new ServiceValidationException("Negative or zero arguments are not accepted");
+            }
+        } catch (DAOException | ServiceValidationException e) {
             throw new ServiceException(e);
         }
     }
